requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
beautifulsoup4==4.10.0
    # via -r requirements.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.10
    # via requests
idna==3.3
    # via requests
python-dotenv==0.19.2
    # via -r requirements.in
requests==2.27.1
    # via -r requirements.in
soupsieve==2.3.1
    # via beautifulsoup4
urllib3==1.26.8
    # via requests
